requirements.txt
absl-py==1.4.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
google-auth==2.18.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
idna==3.4
imageio==2.28.1
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
networkx==3.1
nibabel==5.1.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.10.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==1.12.1+cu113
torch-summary==1.4.5
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.62.3
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.4
